I made simple logic that makes h3 at top work like buttons
