+++ src/components/menu.js
@@ -1,0 +1,6 @@
+import * as methods from "./methods"
+import text from "./text";
+
+export default function myMenu() {
+    
+}
