+++ src/components/text.js
@@ -1,0 +1,9 @@
+const text = {
+  MainInfo:
+  "We are restaurant with long lasting tradition, we are famous for beautiful place, forest, mountains, rivers are us. For many years we bring best possible to your plates and with good prices, we restaurant One Time are one of most enjoyable things you on your little travelling adventure can have. ",
+  timeHeader: "Running hours",
+  place: "You can find us at ...",
+  placeText: "123 NoPlace street TotallyRealPlace",
+};
+
+export default text;
