+++ src/style.css
@@ -4,12 +4,13 @@
   box-sizing: border-box;
 }
 
-@import url('https://fonts.googleapis.com/css2?family=Dancing+Script:wght@700&family=Roboto:wght@300;400;500;700;900&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Dancing+Script:wght@700&family=Roboto:wght@300;400;500;700;900&display=swap");
 
 :root {
-    --main-color: #22c55e;
-    --side-color: #bbf7d095;
-    --border-black-color: rgba(0, 0, 0, 0.54);}
+  --main-color: #22c55e;
+  --side-color: #bbf7d095;
+  --border-black-color: rgba(0, 0, 0, 0.54);
+}
 
 html {
   width: 100%;
@@ -44,8 +45,8 @@ nav > h4 {
 }
 
 nav > h4:hover {
-    color: var(--main-color);
-    cursor: pointer;
+  color: var(--main-color);
+  cursor: pointer;
 }
 
 .logo-container {
@@ -79,9 +80,9 @@ main {
 }
 
 .h2-info-content {
-    text-align: center;
-    font-family: 'Dancing Script', cursive;
-    font-size: 55px;
+  text-align: center;
+  font-family: "Dancing Script", cursive;
+  font-size: 55px;
 }
 
 .info-content {
@@ -91,12 +92,16 @@ main {
 }
 
 .dividing-line {
-    width: 100%;
-    height: 8px;  
-    background-image: repeating-linear-gradient(to right,
-    #045c8c 0px, #c4e17f 50px,
-    #f7fdca 50px, #f7fdca 100px,
-    #e95e54 100px, #fad071 150px,
+  width: 100%;
+  height: 8px;
+  background-image: repeating-linear-gradient(
+    to right,
+    #045c8c 0px,
+    #c4e17f 50px,
+    #f7fdca 50px,
+    #f7fdca 100px,
+    #e95e54 100px,
+    #fad071 150px,
     #09cdb0 200px,
     #5c7cf9 100px,
     #32135a 200px,
@@ -104,9 +109,10 @@ main {
     #c17994 300px,
     #6599e2 300px,
     #045c8c 300px,
-    #909849 200px);
-    background-size: 100% 10px;
-    background-repeat:no-repeat;
+    #909849 200px
+  );
+  background-size: 100% 10px;
+  background-repeat: no-repeat;
 }
 
 .time-container {
@@ -132,7 +138,7 @@ footer {
 }
 
 .footer-content {
-    display: flex;
-    justify-content: flex-start;
-    border-bottom: 20px solid transparent;
+  display: flex;
+  justify-content: flex-start;
+  border-bottom: 20px solid transparent;
 }
